mockResponse as a fixture
